requirements.txt
Flask
psycopg2-binary
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
